Added a new activity which shows rooms queried from the database. The view  looks pretty bad at the moment but it works.
DB should be updated to have coordinates for each room so that we can direct away from it.

The view is set up so that it can accept any building and query accordingly. Currently leaving the view then entering it again breaks the application. Most likely because the table view must be cleared and re-established each time